feat: better ai assistant colors + errors + and inline assist
- error.border and error.background colors helped in making AI Assistant
in AI panel look better when it's showing a stop button. + Tooltips that
are error are more visible. - info.border and info.background are
amazing for inline assistant. I can clearly see what changed. Also it
changed the color of "what the AI changed" essentially to bluer. The
border also communicate which parts I highlighted for the AI Assistant.